Finish almost all of rocket state machine.

+++ variants/units/hera/unit_config.hpp
@@ -58,9 +58,9 @@ const float ANGACC_Z_KI = 0.0;
 const float ANGACC_Z_KD = 0.0;
 
 // Pin configuration
-const int PIN_MAIN_CH   = 1;   // PA5
-const int PIN_DROGUE_CH = 2;   // PC4
-const int PIN_EXT_TEMP_THERM_CH = 1;   // PC1
+const uint8_t PIN_MAIN_CH   = 1;   // PA5
+const uint8_t PIN_DROGUE_CH = 2;   // PC4
+const uint8_t PIN_EXT_TEMP_THERM_CH = 1;   // PC1
 
 }
